+++ lib/reverse_parameters/core_ext.rb
@@ -1,0 +1,2 @@
+require "reverse_parameters/core_ext/method"
+require "reverse_parameters/core_ext/unbound_method"