+++ lib/reverse_parameters/core_ext/method.rb
@@ -1,0 +1,6 @@
+class Method
+  # @return [ReverseParameters]
+  def reverse_parameters
+    ReverseParameters.new(parameters)
+  end
+end