+++ lib/reverse_parameters/core_ext/unbound_method.rb
@@ -1,0 +1,6 @@
+class UnboundMethod
+  # @return [ReverseParameters]
+  def reverse_parameters
+    ReverseParameters.new(parameters)
+  end
+end