+++ spec/reverse_parameters_spec.rb
@@ -247,7 +247,7 @@ describe ReverseParameters do
       it { expect(subject).to eq("&blk") }
 
       it 'blocks as values' do
-        expect(described_class.new(method_proc, blocks_as_values: true).arguments.to_s).to eq("blk")
+        expect(described_class.new(method_proc).arguments(blocks_as_values: true).to_s).to eq("blk")
       end
     end